requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
azure-functions==1.23.0
azure-functions-durable==1.3.2
azure-functions-worker==1.1.9
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
furl==2.1.4
grpcio==1.33.2
grpcio-tools==1.33.2
idna==3.10
importlib_metadata==8.7.0
MarkupSafe==3.0.2
multidict==6.5.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orderedmultidict==1.0.1
propcache==0.3.2
protobuf==3.20.3
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
